agent_zip/requirements.txt
bedrock-agentcore
boto3
pyyaml
